cleanup   move AST.hs to Comments.hs   move HLindent.hs to Refactoring.hs   move Merge.hs to Refactoring.hs   update Main.hs for master changes
